Expose Build options publically
When trying to use the windows version of the socket, I could not build
with the options.

This was an error where only the *Nix version had public fields to edit.

To fix this, the build options has been promoted to the root of the
crate, so it is used on both systems and they end up with the same
config.

This commit also adds an example on the mdns test to ensure it compiles
using the public methods available on all platforms.

It also changes the defaults of build options to listen to multicast
messages, as it is very common to have multiple processes listening on
the same pc.
