+++ src/lib.rs
@@ -1,3 +1,5 @@
+use std::time::Duration;
+
 #[cfg(windows)]
 mod win;
 #[cfg(windows)]
@@ -7,3 +9,19 @@ pub use win::*;
 mod unix;
 #[cfg(not(windows))]
 pub use unix::*;
+
+pub struct MulticastOptions {
+    pub read_timeout: Duration,
+    pub loopback: bool,
+    pub buffer_size: usize,
+}
+
+impl Default for MulticastOptions {
+    fn default() -> Self {
+        MulticastOptions {
+            read_timeout: Duration::from_secs(1),
+            loopback: true,
+            buffer_size: 512,
+        }
+    }
+}
